2022/01: Rework parsing to use anyhow::Result
This is a bit messy, but kinda makes sense and isn't too bad.

My trait impl is (correctly!) using ParseIntError in its Result type, so
it makes sense for a conversion up to anyhow::Error to need to happen.

I think there should be a better way to implement that conversion, but I
haven't yet done so as I'm just cleaning up basic issues.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>